Refactor the LaunchpadTools class into it's own module
